pyproject.toml
[project]
name = "website-extractor"
version = "0.1.0"
description = "Scrape website content"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "ruff>=0.15.1",
    "certifi>=2026.1.4",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
    "django>=5.2.11",
    "bson>=0.5.10",
    "dotenv>=0.9.9",
    "gunicorn>=25.1.0",
    "whitenoise>=6.11.0",
    "openai>=2.30.0",
    "numpy>=2.2.6",
    "djangorestframework>=3.17.1",
    "psycopg[binary]>=3.3.3",
    "tenacity>=9.1.4",
    "pgvector>=0.4.2",
    "tiktoken>=0.12.0",
    "markdown-it-py>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-django>=4.12.0",
    "pyyaml>=6.0.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
testpaths = ["scrapes/tests"]
